Allow for floats

+++ main.py
@@ -5,8 +5,8 @@ while True:
 
     function = input("Which function would you like to carry out? Please enter addition, subtraction, multiplication or division")
 
-    num1 = int(input("Enter number 1"))
-    num2 = int(input("Enter number 2"))
+    num1 = float(input("Enter number 1"))
+    num2 = float(input("Enter number 2"))
 
     if function == "addition":
         print("The answer is " , (num1 + num2))
